% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendtg_check.R
\name{sendtg_check}
\alias{sendtg_check}
\title{Checking the Telegram API Setup in R}
\usage{
sendtg_check()
}
\value{
returns a list with `id`, `is_bot`, `first_name`, and `username`
}
\description{
This function checks whether the setup with your .REnviron was done correctly.
It uses the `telegram.bot::getMe()` to return the relevant information.
If the setup was done incorrectly, it will return an error. Most likely a token issue.
}
\examples{
\dontrun{
sendtg_check()
}


}
