% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendtg_msg.R
\name{sendtg_msg}
\alias{sendtg_msg}
\title{Send Message via Telegram}
\usage{
sendtg_msg(text = NULL, ...)
}
\arguments{
\item{text}{a single \code{character} string containing your message}

\item{...}{other parameters for the \code{sendMessage} function}
}
\value{
a message is sent to the chat ID contained in the .Renviron file
}
\description{
This function sends an file attachment to your chat ID via the \code{telegram.bot} package.
}
\examples{
\dontrun{
sendtg_msg("hello world")
}

}
