% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendtg_setup.R
\name{sendtg_setup}
\alias{sendtg_setup}
\title{Getting Started with Telegram API in R Setup}
\usage{
sendtg_setup()
}
\value{
opens .Renviron file and prints instructions
}
\description{
This function opens the .Renviron file and outlines the steps needed to take to insert the appropriate Telegram API token.
}
\examples{
\dontrun{
sendtg_setup()
}

}
