% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_send_file.R
\name{tg_send_file}
\alias{tg_send_file}
\title{Send File via Telegram}
\usage{
tg_send_file(fpath = NULL, ...)
}
\arguments{
\item{fpath}{a single \code{character} string containing the file fpath}

\item{...}{other arguments fo the \code{sendDocument} function}
}
\value{
a message is sent to the chat ID contained in the .Renviron file with the file
}
\description{
Send File via Telegram
}
