% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_send_img.R
\name{tg_send_img}
\alias{tg_send_img}
\title{Send Images via Telegram}
\usage{
tg_send_img(img = NULL, ...)
}
\arguments{
\item{img}{a \code{character} string containing the \code{path} for your image. URLs can also be used}

\item{...}{other parameters for the \code{sendPhoto} function}
}
\value{
sends an image to your chat ID as an attachment
}
\description{
This function sends a message to your chat ID via the \code{telegram.bot} package.
}
