% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_send_msg.R
\name{tg_send_msg}
\alias{tg_send_msg}
\title{Send Message via Telegram}
\usage{
tg_send_msg(text = NULL, ...)
}
\arguments{
\item{text}{a single \code{character} string containing your message}

\item{...}{other parameters for the \code{sendMessage} function}
}
\value{
a message is sent to the chat ID contained in the .Renviron file
}
\description{
This function sends an file attachment to your chat ID via the \code{telegram.bot} package.
}
