% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendtg_file.R
\name{sendtg_file}
\alias{sendtg_file}
\title{Send File via Telegram}
\usage{
sendtg_file(path = NULL, ...)
}
\arguments{
\item{path}{a single \code{character} string containing the file fpath}

\item{...}{other arguments fo the \code{sendDocument} function}
}
\value{
a message is sent to the chat ID contained in the .Renviron file with the file
}
\description{
Send File via Telegram
}
\examples{
\dontrun{
send_file("important_output.csv")
}

}
